add translate config bump dependencies versions
