+++ package.json
@@ -16,15 +16,17 @@
    },
 
    "dependencies": {
-      "@angular/common": "4.0.1",
-      "@angular/compiler": "4.0.1",
-      "@angular/core": "4.0.1",
-      "@angular/forms": "4.0.1",
-      "@angular/http": "4.0.1",
-      "@angular/platform-browser": "4.0.1",
-      "@angular/router": "4.0.1",
-      "@stratio/egeo": "1.3.3",
-      "@stratio/egeo-theme": "0.5.0",
+      "@angular/common": "4.0.2",
+      "@angular/compiler": "4.0.2",
+      "@angular/core": "4.0.2",
+      "@angular/forms": "4.0.2",
+      "@angular/http": "4.0.2",
+      "@angular/platform-browser": "4.0.2",
+      "@angular/router": "4.0.2",
+      "@ngx-translate/core": "6.0.1",
+      "@ngx-translate/http-loader": "0.0.3",
+      "@stratio/egeo": "1.4.1",
+      "@stratio/egeo-theme": "0.6.0",
       "@stratio/egeo-ui-base": "1.3.0",
       "core-js": "2.4.1",
       "rxjs": "5.3.0",
@@ -32,8 +34,8 @@
    },
 
    "devDependencies": {
-      "@angular/compiler-cli": "4.0.1",
-      "@angular/platform-browser-dynamic": "4.0.1",
+      "@angular/compiler-cli": "4.0.2",
+      "@angular/platform-browser-dynamic": "4.0.2",
       "@ngtools/webpack": "1.3.0",
       "copy-webpack-plugin": "4.0.1",
       "cross-env": "4.0.0",
@@ -47,7 +49,7 @@
       "sass-loader": "6.0.3",
       "style-loader": "0.16.1",
       "typescript": "2.2.2",
-      "webpack": "2.3.3",
+      "webpack": "2.4.1",
       "webpack-dev-server": "2.4.2"
    }
 }
