+++ src/app/modules/errors/errors.module.ts
@@ -17,7 +17,7 @@
 import { NgModule } from '@angular/core';
 
 import { errorsComponents } from '@app/errors';
-// import { SharedModule } from '@starter/shared';
+// import { SharedModule } from '@app/shared';
 
 @NgModule({
    declarations: [
